requirements.txt
# Core dependencies
python-dotenv>=1.0.0
streamlit>=1.28.0

# AI/ML
google-generativeai>=0.5.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# Data processing
pandas>=2.0.0
pypdf>=4.0.0

# Utilities
requests>=2.31.0
